requirements.txt
alembic==1.7.5
beautifulsoup4==4.10.0
blinker==1.4
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
dataclasses==0.8
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
Pillow==8.4.0
pip-search==0.0.7
py-emails==1.2.1
Pygments==2.10.0
requests==2.26.0
rich==10.16.1
soupsieve==2.3.1
SQLAlchemy==1.4.28
typing-extensions==4.0.1
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
WTForms==2.1
zipp==3.6.0
